requirements.txt
Django==1.10.5
gpiozero==1.2.0
djangorestframework==3.5.3
django-filter==1.0.1
markdown==2.6.7
coreapi==2.1.1
django-guardian==1.4.6
gunicorn==19.3.0
psycopg2==2.6.2